Got subsecond timer to work correctly-ish to 10000'ths of a second.  Currently printing time out of ISR so it crashes unless you power on (instead of just resetting)

+++ src/bsp/bsp_shared/time.h
@@ -21,7 +21,7 @@
 typedef struct
 {
    RTC_TimeTypeDef   hour_min_sec;
-   uint16_t          sub_sec;
+   uint32_t          sub_sec;
 }t_Time;
 
 /**
